opt: SEQUOIADBMAINSTREAM-4284 事务rc代码重构优化
git-svn-id: http://192.168.20.11/sequoiadb/trunk@39920 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/coord/coordAuthOperator.cpp
@@ -81,7 +81,7 @@ namespace engine
 
    void _coordAuthOperator::_onSucReply( const MsgOpReply *pReply )
    {
-      if ( pReply->header.messageLength > sizeof( MsgOpReply ) )
+      if ( pReply->header.messageLength > (INT32)sizeof( MsgOpReply ) )
       {
          try
          {
